qq_bot/requirements.txt
nonebot2[fastapi,httpx,websockets]
nonebot-adapter-qq
httpx
